requirements.txt
numpy
pyOpenGL
pyGLM
pyglfw
Pillow